requirements.txt
numpy>=1.26.0
pandas>=2.0.0
scipy>=1.11.0
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0
websockets>=12.0
python-dotenv>=1.0.0
supabase>=2.0.0
aiosqlite>=0.20.0
psycopg2-binary>=2.9.0
